requirements.txt
bidict
absl-py
ray
ray[rllib]
stable_baselines
tensorflow==1.15.4
tensorflow-gpu==1.15.4
sonnet
gfootball